# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ESD Ticket Booking Platform API
  description: |
    Microservices-based event ticket booking platform for IS213.

    All responses follow a standard envelope:
    - **Success**: `{"code": N, "data": {...}}`
    - **Error**: `{"code": N, "message": "..."}`

    All endpoints accept an optional `X-Correlation-ID` header (UUID). If not provided, one is auto-generated for distributed tracing.
  version: 1.0.0
  contact:
    name: ESD Team
servers:
  - url: http://localhost:8000
    description: Kong API Gateway (local development)
...
paths:
...
  /api/payments/create:
    post:
      tags: [Payments]
      summary: Create a Stripe PaymentIntent
      description: Creates a Stripe PaymentIntent in SGD currency.
      operationId: createPayment
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: PaymentIntent created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CreatePaymentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/payments/verify:
    post:
      tags: [Payments]
      summary: Verify a payment
      description: Verifies a Stripe PaymentIntent status and returns transaction details.
      operationId: verifyPayment
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Payment verification result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/VerifyPaymentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/payments/transaction/{booking_id}:
    get:
      tags: [Payments]
      summary: Get transaction by booking ID
      operationId: getTransaction
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
        - name: booking_id
          in: path
          required: true
          schema:
            type: integer
          description: Booking ID
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─────────────────────────────────────────────
  # Notifications
  # ─────────────────────────────────────────────
...
components:
  parameters:
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Correlation ID for distributed tracing. Auto-generated if not provided.
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [code]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 200
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"

    # ── Event ───────────────────────────────────
...
    CreatePaymentRequest:
      type: object
      required: [booking_id, user_id, amount]
      properties:
        booking_id:
          type: integer
          example: 101
        user_id:
          type: string
          example: "firebase_uid_abc123"
        amount:
          type: number
          format: double
          description: Amount in SGD
          example: 128.00
    CreatePaymentResponse:
      type: object
      properties:
        payment_intent_id:
          type: string
          example: "pi_3abc123def456"
        client_secret:
          type: string
          example: "pi_3abc123def456_secret_xyz"
        amount:
          type: number
          format: double
          example: 128.00
    VerifyPaymentRequest:
      type: object
      required: [payment_intent_id]
      properties:
        payment_intent_id:
          type: string
          example: "pi_3abc123def456"
    VerifyPaymentResponse:
      type: object
      properties:
        status:
          type: string
          example: "succeeded"
        booking_id:
          type: integer
          example: 101
        transaction_id:
          type: integer
          example: 55
    Transaction:
      type: object
      properties:
        transaction_id:
          type: integer
          example: 55
        booking_id:
          type: integer
          example: 101
        user_id:
          type: string
          example: "firebase_uid_abc123"
        amount:
          type: number
          format: double
          example: 128.00
        currency:
          type: string
          example: "SGD"
        stripe_payment_intent_id:
          type: string
          example: "pi_3abc123def456"
        status:
          type: string
          enum: [pending, succeeded, failed, refunded, refund_failed]
          example: "succeeded"
        refund_amount:
          type: number
          format: double
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Notification ────────────────────────────
...